+++ packages/client/src/openapi/bracket-notation.test.ts
