+++ packages/client/src/openapi/bracket-notation.ts
