+++ packages/client/src/openapi/serializer.ts
@@ -1,21 +1,21 @@
 import type { JsonValue } from '@orpc/server-standard'
-import type { PublicJSONSerializer } from '../../json-serializer'
+import type { PublicOpenAPIJsonSerializer } from './json-serializer'
 import { ORPCError, toORPCError } from '@orpc/client'
 import { mapEventIterator } from '@orpc/contract'
 import { ErrorEvent, isAsyncIteratorObject } from '@orpc/server-standard'
 import { findDeepMatches } from '@orpc/shared'
-import { JSONSerializer } from '../../json-serializer'
 import * as BracketNotation from './bracket-notation'
+import { OpenAPIJsonSerializer } from './json-serializer'
 
 export interface OpenAPISerializerOptions {
-  jsonSerializer?: PublicJSONSerializer
+  jsonSerializer?: PublicOpenAPIJsonSerializer
 }
 
 export class OpenAPISerializer {
-  private readonly jsonSerializer: PublicJSONSerializer
+  private readonly jsonSerializer: PublicOpenAPIJsonSerializer
 
   constructor(options?: OpenAPISerializerOptions) {
-    this.jsonSerializer = options?.jsonSerializer ?? new JSONSerializer()
+    this.jsonSerializer = options?.jsonSerializer ?? new OpenAPIJsonSerializer()
   }
 
   serialize(data: unknown): unknown {
