+++ packages/openapi/src/adapters/standard/index.ts
@@ -1,5 +1,3 @@
-export * as BracketNotation from './bracket-notation'
 export * from './openapi-codec'
 export * from './openapi-handler'
 export * from './openapi-matcher'
-export * from './openapi-serializer'
