+++ packages/openapi/src/adapters/standard/openapi-codec.ts
@@ -2,9 +2,9 @@ import type { ORPCError } from '@orpc/client'
 import type { AnyProcedure } from '@orpc/server'
 import type { StandardBody, StandardHeaders, StandardRequest, StandardResponse } from '@orpc/server-standard'
 import type { StandardCodec, StandardParams } from '@orpc/server/standard'
+import { OpenAPISerializer } from '@orpc/client/openapi'
 import { fallbackContractConfig } from '@orpc/contract'
 import { isObject } from '@orpc/shared'
-import { OpenAPISerializer } from './openapi-serializer'
 
 export interface OpenAPICodecOptions {
   serializer?: OpenAPISerializer
