+++ packages/openapi/tsconfig.json
@@ -6,7 +6,7 @@
     { "path": "../server" },
     { "path": "../shared" }
   ],
-  "include": ["src"],
+  "include": ["src", "../client/src/openapi/json-serializer.ts"],
   "exclude": [
     "**/*.bench.*",
     "**/*.test.*",
