+++ packages/server/src/adapters/standard/index.ts
@@ -2,5 +2,4 @@ export * from './handler'
 export * from './rpc-codec'
 export * from './rpc-handler'
 export * from './rpc-matcher'
-export * from './rpc-serializer'
 export * from './types'
