+++ packages/vue-colada/package.json
@@ -46,12 +46,11 @@
     "type:check": "tsc -b"
   },
   "peerDependencies": {
-    "@orpc/client": "workspace:*",
     "@pinia/colada": "^0.13.5",
     "vue": ">=3.3.0"
   },
   "dependencies": {
-    "@orpc/server": "workspace:*",
+    "@orpc/client": "workspace:*",
     "@orpc/shared": "workspace:*"
   },
   "devDependencies": {
